requirements.txt
flask
twilio
numpy
tensorflow
pillow
requests
gunicorn